+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to
 
 ### Added
 
+- Enqueues many work orders retries in the same transaction per Oban job.
+  [#2363](https://github.com/OpenFn/lightning/pull/2363)
+
 ### Changed
 
 ### Fixed